pyproject.toml
[tool.poetry]
name = "videodownload"
version = "0.1.0"
description = ""
authors = ["dxsooo <zeohan.dxs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
yt-dlp = "==2022.7.18"
bilix = "^0.6.13"
python-dotenv = {version = "^0.20.0", optional = true}
celery = {version = "^5.2.7", optional = true}
pymongo = {version = "^4.2.0", optional = true}

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[tool.poetry.extras]
celery = ["python-dotenv","celery", "pymongo"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
